improve videoroom example
